refactor: extract injection to separate module
